Remove dependency on five.grok
Simplify package structure and add upgrade steps to deal with changes.
