+++ CHANGES.rst
@@ -6,7 +6,10 @@ There's a frood who really knows where his towel is.
 1.0a2 (unreleased)
 ------------------
 
-- New template with images pager and more layout control. (closes `#4`_).
+- Remove dependency on five.grok (closes `#7`_).
+  [hvelarde]
+
+- Update template with carousel pager and more layout control (closes `#4`_).
   [rodfersou]
 
 - Drop support for Plone 4.2.
@@ -32,3 +35,4 @@ There's a frood who really knows where his towel is.
 
 .. _`#1`: https://github.com/simplesconsultoria/sc.photogallery/issues/1
 .. _`#4`: https://github.com/simplesconsultoria/sc.photogallery/issues/4
+.. _`#7`: https://github.com/simplesconsultoria/sc.photogallery/issues/7
