+++ src/sc/photogallery/browser/configure.zcml
@@ -1,6 +1,18 @@
 <configure
     xmlns="http://namespaces.zope.org/zope"
+    xmlns:browser="http://namespaces.zope.org/browser"
     xmlns:i18n="http://namespaces.zope.org/i18n"
     i18n_domain="sc.photogallery">
+
+  <browser:resourceDirectory name="sc.photogallery" directory="static" />
+
+  <browser:page
+      for="sc.photogallery.content.PhotoGallery"
+      name="view"
+      class="sc.photogallery.browser.View"
+      template="templates/view.pt"
+      permission="zope2.View"
+      layer="sc.photogallery.interfaces.IBrowserLayer"
+      />
 
 </configure>
