+++ src/sc/photogallery/browser/static/photogallery.css
