+++ src/sc/photogallery/browser/static/photogallery.js
