+++ src/sc/photogallery/browser/static/photogallery_icon.png
