+++ src/sc/photogallery/upgrades/v1001/profile/types/Photo_Gallery.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0"?>
+<object name="Photo Gallery">
+  <property name="klass">sc.photogallery.content.PhotoGallery</property>
+</object>
